#1954: split more executor tests; make cancel task run before kill all subprocess; make sure secret in CommandLine argument not log into log
