+++ common/src/com/thoughtworks/go/buildsession/ErrorCommandExecutor.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2016 ThoughtWorks, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.thoughtworks.go.buildsession;
+
+import com.thoughtworks.go.domain.BuildCommand;
+
+// this is not part of the protocol. only for error flow testing
+public class ErrorCommandExecutor implements BuildCommandExecutor {
+    @Override
+    public boolean execute(BuildCommand command, BuildSession buildSession) {
+        throw new RuntimeException(command.getStringArg("message"));
+    }
+}
