+++ common/test/unit/com/thoughtworks/go/buildsession/MkdirsCommandExecutorTest.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2016 ThoughtWorks, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.thoughtworks.go.buildsession;
+
+import org.junit.Test;
+
+import java.io.File;
+
+import static com.thoughtworks.go.domain.BuildCommand.mkdirs;
+import static com.thoughtworks.go.domain.BuildCommand.test;
+import static com.thoughtworks.go.domain.JobResult.Failed;
+import static com.thoughtworks.go.domain.JobResult.Passed;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+
+public class MkdirsCommandExecutorTest extends BuildSessionBasedTestCase {
+    @Test
+    public void mkdirsShouldCreateDirectoryIfNotExists() {
+        runBuild(mkdirs("foo"), Passed);
+        assertThat(new File(sandbox, "foo").isDirectory(), is(true));
+    }
+
+    @Test
+    public void mkdirsShouldFailIfDirExists() {
+        runBuild(mkdirs("foo"), Passed);
+        runBuild(mkdirs("foo"), Failed);
+        assertThat(new File(sandbox, "foo").isDirectory(), is(true));
+    }
+
+    @Test
+    public void testDirectoryExistsBeforeMkdir() {
+        File dir = new File(sandbox, "foo");
+        runBuild(mkdirs("foo"), Passed);
+        runBuild(mkdirs("foo").setTest(test("-nd", dir.getPath())), Passed);
+        assertThat(new File(sandbox, "foo").isDirectory(), is(true));
+    }
+}
